Move generators to separate internal package, run generate for demo script
