+++ gopages/internal/generate/godoc.html.go
@@ -1,4 +1,4 @@
-package main
+package generate
 
 const (
 	godocHTML = `<!DOCTYPE html>
